Add further utility methods to utils
